Move calculateProxyAddress and formatInternalContracts to fork/utils
